homework1

+++ src/Homework1/Homework1_1.java
@@ -4,6 +4,7 @@ public class Homework1_1 {
     public static void main(String[] args) {
         int b = 5;
         int c = 6;
-       float a=4*(b+c-1)/2;
-       System.out.println(a);
-    }}
+        int a = 4 * (b + c - 1) / 2;
+        System.out.println(a);
+    }
+}
